fixed vorticity
The formula for eta in the paper is misleading, the subscript i does not make sense
and should not be there.
Eta is simply the standard SPH gradient over the magnitudes of omega. The standard
gradient calculation can be found in the Monaghan 1992 paper and is used
throughout PBF.
Unfortunately this requires a new kernel as the omega for each particle has to be
calculated before the gradient can be applied.