moves sorting to the beginning
CAUTION: this introduces a VERY weird issue with particles being reset to 0,0,0
Currently this is only happening for Iris Pro 5200 on Apple, not on the proccessor
and not under Windows.

Putting a printf statement anywhere in any kernel seems to "fix" the issue.
The issue does only occur if either applyViscosity or applyVorticity is active.
However if this only works if the kernels are not called at all from the Host,
if the code function bodies for each kernel are simply commented out the issue
STILL occurs.
Another way to "fix" the issue is to normalize the velocity if gets to large,
however only if it is done with the OpenCL normalize() function. Manually dividing
by the length does not work, nor does fast_normalize(). This might point to
numerical issues in Apple's implementation but commenting out the code like described
before still does not make any sense at all.
All in all this really sucks and the real solution is not in sight...